fix(compiler): move semantic memory utilities
Move memory declaration parsing and memory flag helpers under semantic/utils, and update semantic imports directly without compatibility exports.
