+++ packages/compiler/src/semantic/utils/memoryFlags.test.ts
