+++ packages/compiler/src/semantic/utils/memoryFlags.ts
