+++ packages/compiler/src/semantic/utils/memoryInstructionParser.test.ts
